refactor(WebAPI): Introduce AllJobListener and SchedulerListener for job status updates
